ExactOnlineQuery.CreateODataQuery:
AND is added as a query part list element which gets joined to a querystring by &. For example: $filter=City+eq+'AMSTERDAM'&AccountIsSupplier+eq+true.

This results in an incorrect where clause. the correct where clause should be:

$filter=City+eq+'AMSTERDAM'+and+AccountIsSupplier+eq+true

Which gets correctly translated to:

SELECT *  FROM ExactOnline.Addresses WHERE (City = 'Amsterdam') AND (AccountIsSupplier = True)