Fetch only the first page by default
This change will make the API only fetch the first page by default for
the user. This prevents (unknowingly) fetching an unknown amount of records
into memory.

Instead, the API provides a ExactOnlineQuery#GetPages() method which
returns an IEnumerable of the pages of the query. Each page is
represented as a list with at most 60 records.